- Set deployment target from 10.7 to 10.8. - Fix code to suppress warnings after changing deployment target.

+++ Projeta/MainWindowController.m
@@ -136,7 +136,7 @@
     [projectViewController setMainWindowController:self];
     
     // index du tab actuel.
-    int selectedTabIndex = [[[mainWindowViewController projectListViewController] prjTabView] indexOfTabViewItem:[[[mainWindowViewController projectListViewController] prjTabView] selectedTabViewItem]];
+    NSInteger selectedTabIndex = [[[mainWindowViewController projectListViewController] prjTabView] indexOfTabViewItem:[[[mainWindowViewController projectListViewController] prjTabView] selectedTabViewItem]];
